refactor: split DaikinOne into a client subpackage
The 614-line daikinone.py mixed four concerns: domain models, wire-format
primitives, HTTP+auth transport, and payload-to-domain mapping. Split into
a client/ subpackage so each file has one job:

- client/models.py   -- dataclass/enum domain types
- client/wire.py     -- URL constants + DaikinDeviceDataResponse
- client/transport.py -- DaikinTransport: auth state + HTTP
- client/mapping.py  -- pure payload-to-domain functions (one per equipment kind)
- client/client.py   -- DaikinOne orchestration (transport + cache + setters)

The equipment mapping no longer mutates a shared dict from inside a single
function; each equipment kind has its own _map_* that returns None when
absent, composed by map_equipment. All callers (__init__, climate, select,
sensor, entity, config_flow, diagnostics, tests) now import from
custom_components.daikinone.client submodules.
